добавил DTO, доработал менеджеров
