+++ src/Exceptions/BaseException.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace Src\Exceptions;
+
+use Exception;
+
+class BaseException extends Exception
+{
+}
