+++ src/Exceptions/FileSystem/FileExistsException.php
@@ -1,0 +1,15 @@
+<?php
+
+
+namespace Src\Exceptions\FileSystem;
+
+use Src\Exceptions\BaseException;
+
+class FileExistsException extends BaseException
+{
+    public function __construct($message = null)
+    {
+        $this->message = $message ?? 'File already exists';
+        parent::__construct($this->message);
+    }
+}
