+++ src/Exceptions/Task/actions/AccessIsDeniedException.php
@@ -1,0 +1,14 @@
+<?php
+
+
+namespace Src\Exceptions;
+
+
+class AccessIsDeniedException extends BaseException
+{
+    public function __construct($message = null)
+    {
+        $this->message = $message ?? 'Access is denied';
+        parent::__construct($this->message);
+    }
+}
