+++ src/Logic/FileSystem/Data/IDTO.php
@@ -1,0 +1,9 @@
+<?php
+
+
+namespace Logic\FileSystem\Data;
+
+
+interface IDTO
+{
+}
