--- src/Logic/FileSystem/IConvertor.php
@@ -1,8 +1,0 @@
-<?php
-
-namespace Logic\FileSystem;
-
-interface IConvertor
-{
-    public function convertData(string $path);
-}
