+++ src/Logic/FileSystem/Managers/IReader.php
@@ -1,0 +1,15 @@
+<?php
+
+namespace Logic\FileSystem\Managers;
+
+interface IReader
+{
+    public function openFile(string $path, string $mode);
+
+    public function checkExistFile(string $filePath) : bool;
+
+    public function readFile(\SplFileObject $file): array;
+
+    public function getDTO(\SplFileObject $file);
+
+}
