+++ src/Logic/FileSystem/Managers/IWriter.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace Logic\FileSystem\Managers;
+
+interface IWriter
+{
+    public function createFile(string $path);
+
+    public function writeInFile($file, string $content): void;
+
+    public function closeFile($file): void;
+
+    public function getContent($dto): string;
+}
