--- src/Logic/FileSystem/managers/IReader.php
@@ -1,10 +1,0 @@
-<?php
-
-namespace Logic\FileSystem\managers;
-
-interface IReader
-{
-    public function openFile(string $path, string $mode);
-
-    public function readFile(\SplFileObject $file): array;
-}
