--- src/Logic/FileSystem/managers/IWriter.php
@@ -1,12 +1,0 @@
-<?php
-
-namespace Logic\FileSystem\managers;
-
-interface IWriter
-{
-    public function createFile(string $path);
-
-    public function writeInFile($file, string $content): void;
-
-    public function closeFile($file): void;
-}
