--- src/exceptions/BaseException.php
@@ -1,9 +1,0 @@
-<?php
-
-namespace src\exceptions;
-
-use Exception;
-
-class BaseException extends Exception
-{
-}
