--- src/exceptions/fileSystem/FileDoesNotExistsException.php
@@ -1,15 +1,0 @@
-<?php
-
-
-namespace src\exceptions\fileSystem;
-
-use src\exceptions\BaseException;
-
-class FileDoesNotExistsException extends BaseException
-{
-    public function __construct($message = null)
-    {
-        $this->message = $message ?? 'File does not exists';
-        parent::__construct($this->message);
-    }
-}
