--- src/exceptions/task/TaskStatusNotExistException.php
@@ -1,14 +1,0 @@
-<?php
-
-
-namespace src\exceptions;
-
-
-class TaskStatusNotExistException extends BaseException
-{
-    public function __construct($message = null)
-    {
-        $this->message = $message ?? 'Status not exist';
-        parent::__construct($this->message);
-    }
-}
