--- src/exceptions/task/TaskStatusNotHasActionsException.php
@@ -1,14 +1,0 @@
-<?php
-
-
-namespace src\exceptions;
-
-
-class TaskStatusNotHasActionsException extends BaseException
-{
-    public function __construct($message = null)
-    {
-        $this->message = $message ?? 'Status don\'t have actions';
-        parent::__construct($this->message);
-    }
-}
