--- src/tests/TestActionStart.php
@@ -1,36 +1,0 @@
-<?php
-
-
-namespace src\tests;
-
-
-use Logic\Task;
-use src\Logic\actions\TaskActionStart;
-
-class TestActionStart
-{
-    public function testIsHasStart() : bool
-    {
-        $task = new Task(1,2);
-        $action = new TaskActionStart();
-        return assert(in_array($action, $task->getActionForStatus(Task::STATUS_NEW)),  'TaskAction не разрешен');
-    }
-
-    public function testStatusAfterStart() : bool
-    {
-        $task = new Task(1,2);
-        $action = new TaskActionStart();
-        $status = Task::STATUS_ACTIVE;
-        return assert($task->getNextStatus($action) === $status,
-            $task->getNextStatus($action) . ' != ' .  $status . ' |  статус после выполнения ' . $action::getInnerName()
-            . ' не соответствует этому действию' );
-    }
-
-    public function testStartActionCheckRight() : bool
-    {
-        $action = new TaskActionStart();
-        $test = $action->checkRights(1, 2, 2);
-        return assert($test, $action::getInnerName() . ' действие не доступно для пользователя');
-    }
-
-}
